+++ actingweb/interface/__init__.py
@@ -10,7 +10,7 @@ from .actor_interface import ActorInterface
 from .property_store import PropertyStore
 from .trust_manager import TrustManager
 from .subscription_manager import SubscriptionManager
-from .hooks import HookRegistry, property_hook, callback_hook, subscription_hook
+from .hooks import HookRegistry, property_hook, callback_hook, app_callback_hook, subscription_hook, method_hook, action_hook
 
 __all__ = [
     "ActingWebApp",
@@ -21,5 +21,8 @@ __all__ = [
     "HookRegistry",
     "property_hook",
     "callback_hook", 
+    "app_callback_hook",
     "subscription_hook",
+    "method_hook",
+    "action_hook",
 ]